+++ project/src/store/app-slice/selectors/get-app-status/get-app-status.ts
