+++ project/src/store/app-slice/selectors/get-city/get-city.ts
