+++ project/src/store/data-slice/selectors/get-current-offer/get-current-offer.ts
