+++ project/src/store/data-slice/selectors/get-near-offers/get-near-offers.ts
