+++ project/src/store/data-slice/selectors/get-offers/get-offers.ts
