+++ project/src/store/index.ts
@@ -1,5 +1,5 @@
 export {default as store} from '../store/store';
 export {loadOffers, checkLogin, login, logout, loadCurrentOffer, postReview} from './actions';
-export {setCity} from './app-process/app-process';
-export {resetCurrentOffer} from './data-process/data-process';
+export {setCity} from './app-slice/app-slice';
+export {resetCurrentOffer} from './data-slice/data-slice';
