+++ project/src/store/reducer.ts
@@ -1,11 +1,11 @@
 import {combineReducers} from '@reduxjs/toolkit';
 import {NameSpace} from '../const';
-import {dataProcess} from './data-process/data-process';
-import {appProcess} from './app-process/app-process';
-import {userProcess} from './user-process/user-process';
+import {dataSlice} from './data-slice/data-slice';
+import {appSlice} from './app-slice/app-slice';
+import {userSlice} from './user-slice/user-slice';
 
 export const reducer = combineReducers({
-  [NameSpace.Data]: dataProcess.reducer,
-  [NameSpace.App]: appProcess.reducer,
-  [NameSpace.User]: userProcess.reducer,
+  [NameSpace.Data]: dataSlice.reducer,
+  [NameSpace.App]: appSlice.reducer,
+  [NameSpace.User]: userSlice.reducer,
 });
