+++ project/src/store/selectors.ts
@@ -1,8 +1,8 @@
-export {default as getCity} from './app-process/selectors/get-city/get-city';
-export {default as getAppStatus} from './app-process/selectors/get-app-status/get-app-status';
-export {default as getUser} from './user-process/selectors/get-user/get-user';
-export {default as getAuthStatus} from './user-process/selectors/get-auth-status/get-auth-status';
-export {default as getOffers} from './data-process/selectors/get-offers/get-offers';
-export {default as getCurrentOffer} from './data-process/selectors/get-current-offer/get-current-offer';
-export {default as getNearOffers} from './data-process/selectors/get-near-offers/get-near-offers';
-export {default as getReviews} from './data-process/selectors/get-reviews/get-reviews';
+export {default as getCity} from './app-slice/selectors/get-city/get-city';
+export {default as getAppStatus} from './app-slice/selectors/get-app-status/get-app-status';
+export {default as getUser} from './user-slice/selectors/get-user/get-user';
+export {default as getAuthStatus} from './user-slice/selectors/get-auth-status/get-auth-status';
+export {default as getOffers} from './data-slice/selectors/get-offers/get-offers';
+export {default as getCurrentOffer} from './data-slice/selectors/get-current-offer/get-current-offer';
+export {default as getNearOffers} from './data-slice/selectors/get-near-offers/get-near-offers';
+export {default as getReviews} from './data-slice/selectors/get-reviews/get-reviews';
