+++ project/src/store/user-slice/selectors/get-auth-status/get-auth-status.ts
