+++ project/src/store/user-slice/selectors/get-user/get-user.ts
